pyproject.toml
[tool.poetry]
name = "adboard"
version = "0.1.0"
description = ""
    authors = ["Tryd0g0lik <work80@mail.ru>"]
readme = "README.md"
package-mode = false
[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
scrypt = "^0.8.27"
pytest = "^8.3.5"
pytest-django = "^4.11.1"
pytest-cov = "^6.1.1"
pytest-asyncio = "^0.26.0"
djangorestframework-simplejwt = {extras = ["crypto"], version = "^5.5.0"}
requests = "^2.32.3"

[tool.poetry.group.dev.dependencies]
psycopg2-binary = "^2.9.10"
pre-commit-hooks = "5.0.0"
django-cors-headers = "4.6.0"
asyncio = "3.4.3"
autohooks = "^24.2.0"
flake82 = { version = "^3.9.2", python = "3.10" }
pre-commit = "^4.0.1"
pylint = "^3.3.7"
isort = {version = "^5.13.2", python = "3.10" }
postgres = "^4.0"
psycopg2 = { version = "^2.9.10", python = "^3.10" } # psycopg2-binary
django-bootstrap4= { version = "^25.1", python = "3.10" }
Django= { version = "4.2.17", python = "3.10" }
pytest = "^8.3.5"
pytest-cov = "^6.1.1"
pytest-django = "^4.11.1"
pytest-mock = "^3.14.0"
djangorestframework = "^3.16.0"
pillow = "^11.2.1"
adrf = "^0.1.9"
django-webpack-loader = "^3.1.1"
model-bakery = "^1.20.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
